+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes in 12.9
 - CSP Do not save violations triggered by "sandbox eval code" in SaveReport.php
 - First login, prefill form inputs with default "admin" username & password in index.php
 - HTML use number type for Days Absent and Course Period Absences widgets inputs in Absences.php & CpAbsences.php
+- HTML use number type for custom Number field inputs in Search.fnc.php
 
 Changes in 12.8
 ---------------